+++ changelog.rst
@@ -1,7 +1,14 @@
 Changelog
 =========
 
-Last change: 06-MAY-2025 MTS
+Last change: 15-JUL-2025 MTS
+
+0.8.1
+-----
+- Added ``n_events`` to cluster centers, i.e., number of binding events per cluster
+- .yaml files contain Picasso version number for easier tracking
+- Improved fiducial picking
+- Bug fixes and other cosmetic changes
 
 0.8.0
 -----